Use builder pattern to create configuration
